Create user sqlite component and move related logic to it
