+++ clock/storage/data_source/data_sources/sqlite/__init__.py
@@ -1,6 +1,7 @@
 import sqlite3
 
 from clock.storage.data_source.data_source import StorageDataSource
+from clock.storage.data_source.data_sources.sqlite.component.components.user import UserSqliteComponent
 
 
 DATABASE_FILENAME = "state/clock.db"
@@ -12,30 +13,16 @@ class SqliteStorageDataSource(StorageDataSource):
     def __init__(self):
         # initialized in init to avoid creating sqlite objects outside the thread in which it will be operating
         self.connection = None
+        self.user = None  # type: UserSqliteComponent
 
     def init(self):
         self.connection = sqlite3.connect(DATABASE_FILENAME)
         self.connection.row_factory = sqlite3.Row  # improved rows
+        self.user = UserSqliteComponent(self.connection)
+        self.user.init()
         self._init_db()
 
     def _init_db(self):
-        self.__sql("create table if not exists user ("
-                   "user_id integer primary key not null,"
-                   "first_name text,"
-                   "last_name text,"
-                   "username text,"
-                   "language_code text,"
-                   "timestamp_added text"
-                   ")")
-        self.__sql("create table if not exists user_history ("
-                   "user_id integer not null,"
-                   "first_name text,"
-                   "last_name text,"
-                   "username text,"
-                   "language_code text,"
-                   "timestamp_added text,"
-                   "timestamp_removed text"
-                   ")")
         self.__sql("create table if not exists query ("
                    "timestamp text,"
                    "user_id integer not null,"
@@ -57,29 +44,7 @@ class SqliteStorageDataSource(StorageDataSource):
                    ")")
 
     def save_user(self, user_id: int, first_name: str, last_name: str, username: str, language_code: str):
-        first_name = self.__empty_if_none(first_name)
-        last_name = self.__empty_if_none(last_name)
-        username = self.__empty_if_none(username)
-        language_code = self.__empty_if_none(language_code)
-        if not self.__is_user_saved_equal(user_id, first_name, last_name, username, language_code):
-            self.__add_to_user_history(user_id)
-            self.__sql("insert or replace into user "
-                       "(user_id, first_name, last_name, username, language_code, timestamp_added) "
-                       "values (?, ?, ?, ?, ?, strftime('%s', 'now'))",
-                       (user_id, first_name, last_name, username, language_code))
-
-    def __is_user_saved_equal(self, user_id: int, first_name: str, last_name: str, username: str, language_code: str):
-        return self.__sql("select 1 from user where "
-                          "user_id = ? and first_name = ? and last_name = ? and username = ? and language_code = ?",
-                          (user_id, first_name, last_name, username, language_code)).fetchone()
-
-    def __add_to_user_history(self, user_id: int):
-        # if user does not exists in user table, nothing will be inserted into user_history, as expected for new users
-        self.__sql("insert into user_history "
-                   "(user_id, first_name, last_name, username, language_code, timestamp_added, timestamp_removed) "
-                   "select user_id, first_name, last_name, username, language_code, "
-                   "timestamp_added, strftime('%s', 'now') "
-                   "from user where user_id = ?", (user_id,))
+        self.user.save_user(user_id, first_name, last_name, username, language_code)
 
     def save_query(self, user_id: int, timestamp: str, query: str, offset: str, locale: str, results_found_len: int,
                    results_sent_len: int, processing_seconds: float):
@@ -101,7 +66,3 @@ class SqliteStorageDataSource(StorageDataSource):
 
     def __sql(self, sql: str, params=()):
         return self.connection.execute(sql, params)
-
-    @staticmethod
-    def __empty_if_none(field: str):
-        return field if field is not None else ""
